requirements.txt
noise==1.2.2
numpy==1.20.0
Pillow==8.1.0
pygame==2.0.1
